+++ cucumber.js
@@ -1,6 +1,8 @@
 var common = [
   '--require-module babel-register',
-  `--format ${process.env.CI ? 'progress' : 'progress-bar'}`,
+  `--format ${
+    process.env.CI || !process.stdout.isTTY ? 'progress' : 'progress-bar'
+  }`,
   '--format rerun:@rerun.txt',
   '--format usage:usage.txt',
 ].join(' ')