+++ src/formatter/get_color_fns.js
@@ -1,17 +1,41 @@
+import _ from 'lodash'
 import colors from 'colors/safe'
 import Status from '../status'
 
 export default function getColorFns(enabled) {
-  colors.enabled = enabled
-  colors.setTheme({
-    [Status.AMBIGUOUS]: 'red',
-    [Status.FAILED]: 'red',
-    [Status.PASSED]: 'green',
-    [Status.PENDING]: 'yellow',
-    [Status.SKIPPED]: 'cyan',
-    [Status.UNDEFINED]: 'yellow',
-    location: 'grey',
-    tag: 'cyan',
-  })
-  return colors
+  if (enabled) {
+    return {
+      [Status.AMBIGUOUS]: ::colors.red,
+      [Status.FAILED]: ::colors.red,
+      [Status.PASSED]: ::colors.green,
+      [Status.PENDING]: ::colors.yellow,
+      [Status.SKIPPED]: ::colors.cyan,
+      [Status.UNDEFINED]: ::colors.yellow,
+      location: ::colors.gray,
+      tag: ::colors.cyan,
+
+      // For assertion-error-formatter
+      diffAdded: ::colors.green,
+      diffRemoved: ::colors.red,
+      errorMessage: ::colors.red,
+      errorStack: ::colors.gray,
+    }
+  } else {
+    return {
+      [Status.AMBIGUOUS]: _.identity,
+      [Status.FAILED]: _.identity,
+      [Status.PASSED]: _.identity,
+      [Status.PENDING]: _.identity,
+      [Status.SKIPPED]: _.identity,
+      [Status.UNDEFINED]: _.identity,
+      location: _.identity,
+      tag: _.identity,
+
+      // For assertion-error-formatter
+      diffAdded: _.identity,
+      diffRemoved: _.identity,
+      errorMessage: _.identity,
+      errorStack: _.identity,
+    }
+  }
 }